fix: address Copilot review findings on memcached proxy managers
- Propagate permanent CAS/add failures (e.g. VALUE_TOO_LARGE, OUT_OF_MEMORY)
  instead of retrying them forever as ordinary conflicts.
- Reject lock keys that would exceed memcached's 250-byte key limit once the
  ".bucket4j-lock" suffix is appended.
- Extract a shared MemcachedExpirations helper that rounds TTLs up to whole
  seconds and converts values beyond memcached's 30-day cutoff into an
  absolute Unix timestamp, instead of sending a relative value that memcached
  misreads as a 1970 timestamp.
- Store lock expiration as a long to avoid overflow when casting to int.

+++ bucket4j-memcached/src/main/java/io/github/bucket4j/memcached/Bucket4jMemcached.java
@@ -139,12 +139,12 @@ public class Bucket4jMemcached {
 
     public static class MemcachedLockBasedProxyManagerBuilder<K> extends AbstractProxyManagerBuilder<K, MemcachedLockBasedProxyManager<K>, MemcachedLockBasedProxyManagerBuilder<K>> {
 
-        private static final int DEFAULT_LOCK_EXPIRATION_SECONDS = 30;
+        private static final long DEFAULT_LOCK_EXPIRATION_SECONDS = 30;
         private static final long DEFAULT_LOCK_POLL_PERIOD_MILLIS = 50;
 
         private final MemcacheClient<byte[]> client;
         private Mapper<K> keyMapper;
-        private int lockExpirationSeconds = DEFAULT_LOCK_EXPIRATION_SECONDS;
+        private long lockExpirationSeconds = DEFAULT_LOCK_EXPIRATION_SECONDS;
         private long lockPollPeriodMillis = DEFAULT_LOCK_POLL_PERIOD_MILLIS;
 
         private MemcachedLockBasedProxyManagerBuilder(MemcacheClient<byte[]> client, Mapper<K> keyMapper) {
@@ -184,11 +184,11 @@ public class Bucket4jMemcached {
          */
         public MemcachedLockBasedProxyManagerBuilder<K> lockExpiration(Duration lockExpiration) {
             Objects.requireNonNull(lockExpiration);
-            this.lockExpirationSeconds = (int) Math.max(1, lockExpiration.toSeconds());
+            this.lockExpirationSeconds = Math.max(1, lockExpiration.toSeconds());
             return this;
         }
 
-        public int getLockExpirationSeconds() {
+        public long getLockExpirationSeconds() {
             return lockExpirationSeconds;
         }
